+++ src/app/components/dashboard/turno/historial-turnos.component.ts
@@ -1,3 +1,4 @@
+import { EliminarComponent } from './../eliminar/eliminar.component';
 import { Cliente } from './../models/cliente';
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
@@ -36,7 +37,6 @@ export class HistorialTurnosComponent implements OnInit {
 
   ngOnInit(): void {
     this.router.params.subscribe((params: Params) => { this.id = params['id'] })
-
     this.clienteService.findCliente(this.id).subscribe(data => {
       this.cliente = data;
     })
@@ -77,17 +77,26 @@ export class HistorialTurnosComponent implements OnInit {
     this.dialog.open(NuevoModifTurnoComponent, dialogConfig);
   }
 
-  onDelete(id:number){
-    console.log(id);
-    this.turnoService.delete(id).subscribe(
-      data => {
-        this.toastr.success("Turno eliminado con Exito!","OK", { timeOut: 3000 });
-        this.cargarTurnos();
-      },
-      err => {
-        this.toastr.error(err.error.mensaje, 'Error al eliminar el Turno', { timeOut: 3000 });
+  onDelete(id:number, turno: Turno){
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    dialogConfig.width = '250px';
+    dialogConfig.data ="de " + turno.cliente.nombre +" "+ turno.cliente.apellido +" - Fecha "+ turno.fecha_turno;
+    dialogConfig.id = "Turno";
+    this.dialog.open(EliminarComponent, dialogConfig).afterClosed().subscribe(result => {
+      if (result == true) {
+        console.log(id);
+        this.turnoService.delete(id).subscribe(
+          data => {
+            console.log(result);
+            this.toastr.success('Turno Eliminado con exito!', 'OK', { timeOut: 3000 });
+            this.cargarTurnos();
+          })
+      } else {
+        console.log("no queria eliminar");
       }
-    )
+    });
   }
 
 }
